Fixes MULSH opcode and strings decompilation
From page 455 of Xtensa (ISA) Reference Manual the opcode for MULSH is 0xB2 not 0xB1. B1 is SRA
AS_ASCIIC | AS_ASCIIZ flags do not get zeroed out when this dictionary is initialized and therefore strings are not displayed properly. May just be an issue with my version of IDA. Ignore if strings show up correctly on your version.